feat: integra botões de alteração de quantidade à lógica do estoque no index.php

+++ index.php
@@ -80,10 +80,13 @@ require_once "init.php";
                         echo "<td>R$ " . number_format($produto['preco'], 2, ',', '.') . "</td>";
                         echo "<td>R$ " . number_format($produto['quantidade'] * $produto['preco'], 2, ',', '.') . "</td>";
                         echo "<td>
-                                <button class='btn-acao editar'>-1</button>
-                                <button class='btn-acao editar'>+1</button>
-                                <a href='excluir_produto.php?id={$index}' 
-                                    onclick='return confirm(\"Tem certeza que deseja excluir?\")'>
+                                <a href='alterar_quantidade.php?id={$index}&acao=subtrair'>
+                                    <button class='btn-acao editar'>-1</button>
+                                </a>
+                                <a href='alterar_quantidade.php?id={$index}&acao=somar'>
+                                    <button class='btn-acao editar'>+1</button>
+                                </a>
+                                <a href='excluir_produto.php?id={$index}'onclick='return confirm(\"Tem certeza que deseja excluir?\")'>
                                     <button class='btn-acao excluir'>Excluir</button>
                                 </a>
                             </td>";
